Completed stage 5 (abstracted version)
The game can now be played with any words the user chooses to play with. 
Each word will always be "stronger" than one half, and "weaker" than the other half of the list.
If no custom word list is entered, the game is played with the default "rock, paper, scissors".